Building out the draft pack ability further

+++ src/main/java/com/conoftherings/draft/Pack.java
@@ -6,19 +6,19 @@ import com.conoftherings.playercards.Card;
 
 public class Pack {
 
-    private final int id;
+    private final long id;
     private final String userName;
     private final String packName;
     private final List<Card> cards;
 
-    public Pack(int id, String userName, String packName, List<Card> cards) {
+    public Pack(long id, String userName, String packName, List<Card> cards) {
         this.id = id;
         this.userName = userName;
         this.packName = packName;
         this.cards = cards;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
